change channelID from int to long

+++ features/content/play_broadcaster/src/main/java/com/tokopedia/play/broadcaster/domain/repository/PlayBroProductRepository.kt
@@ -30,6 +30,6 @@ interface PlayBroProductRepository {
     suspend fun setProductTags(channelId: String, productIds: List<String>)
 
     suspend fun getProductTagSummarySection(
-        channelID: Int,
+        channelID: Long,
     ): Pair<List<ProductTagSectionUiModel>, Int>
 }
